Have point and plane driver follow folder new org
