+++ src/Defs.cpp
@@ -19,7 +19,6 @@
 *****************************************************************************/
 
 #include "Defs.hpp"
-#include "PointAndPlaneDriver.hpp"
 #include "TriangleSegment.hpp"
 
 #include <iostream>
