+++ src/map-director/ViewGrid.hpp
@@ -161,7 +161,8 @@ template <typename T>
 void ViewGridInserter<T>::advance() {
     using namespace cul::exceptions_abbr;
     if (filled()) {
-        throw RtError{""};
+        throw RtError{"ViewGridInserter::advance: cannot advance a filled "
+                      "inserter"};
     }
     auto el_count = m_elements.size();
     std::get<1>(m_index_pairs(m_position)) = el_count;
