+++ tests/TaskController-tests.cpp
@@ -19,8 +19,8 @@
 *****************************************************************************/
 
 #include "../src/TasksController.hpp"
-// forward declarations
-#include "../src/PointAndPlaneDriver.hpp"
+
+#include "../src/point-and-plane.hpp"
 
 #include <ariajanke/cul/TreeTestSuite.hpp>
